Add macOS/iOS support (via MoltenVK), Take 2 (#8)
* Replace Linux-specific strerror_r() with generic POSIX solution

The version of strerror_r() used in virtualalloc.hpp is Linux-specific.
In order to compile on other POSIX systems, we need to replace it with a
more generic POSIX solution.

* Rename VUDA_LINUX to VUDA_UNIX

VUDA_LINUX is now generic-Unix instead of just Linux,
so hooray for a new name.

* Add support for MoltenVk on MacOS/iOS

After the previous two commits, the only change needed
is to detect Darwin-based systems using the __APPLE__
preprocessor directive.

* Fix for compiling with clang

The way kernelprogram was templated before in logicaldevice.inl was
causing clang to barf. Clang thinks that the template argument was not a
constant expression.